backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
numpy==1.24.3
opencv-python==4.8.0.76
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
